+++ Source/App/Shredder/Helper/ValidationHelper.swift
@@ -3,7 +3,7 @@
 //  Shredder
 //
 //  Created by Arnold Nefkens on 03/10/2018.
-//  Copyright © 2018 Pro Warehouse.
+//  Copyright © 2019 Pro Warehouse.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -168,5 +168,92 @@ class ValidationHelper: NSObject, Logging {
         let needsConnection = flags.contains(.connectionRequired)
 
         return (isReachable && !needsConnection)
+    }
+
+    /// Generates the Arguments to be send to the startosinstall command.
+    ///
+    /// Will search a fixed loction for .pkg files. When found any, they are added to the startosinstall command.
+    /// Will always return always an array with the two minimum arguments to use: --agreetolicense & --eraseinstall
+    ///
+    /// - Returns: Array of Strings, that hold the additional arguments.
+    func fetchStartOSInstallArguments() -> [String] {
+        var argumentsForStartOSInstall: [String] = ["--agreetolicense", "--eraseinstall"]
+        do {
+            let pathForInstallersFolder: String = "/Library/Application Support/EraseInstall/Packages/"
+            let URLFullPath: URL = URL(fileURLWithPath: pathForInstallersFolder)
+            let directoryContents = try FileManager.default.contentsOfDirectory(at: URLFullPath,
+                                                                                includingPropertiesForKeys: nil,
+                                                                                options: [])
+
+            // Filtered on .pkg files
+            let installersFound = directoryContents.filter { $0.pathExtension == "pkg" }
+                                .sorted(by: { $0.lastPathComponent < $1.lastPathComponent })
+            if installersFound.count > 0 {
+                log(message: "Packages found.")
+                for installer in installersFound {
+                    // validate the package
+                    if validatePackage(path: installer.path) {
+                        argumentsForStartOSInstall.append("--installpackage")
+                        argumentsForStartOSInstall.append(installer.path)
+                        log(message: "Package: \(installer.path)")
+                    }
+                }
+            }
+        } catch {
+            log(message: "No Additional Packages found.")
+        }
+
+        return argumentsForStartOSInstall
+    }
+
+    /// Validats the found installer.
+    ///
+    /// - Parameter path: String to the package
+    /// - Returns: True if package is correct.
+    private func validatePackage(path: String) -> Bool {
+        //Fetch URL of Validation Script
+        let pathToPackageValidator = Bundle.main.path(forResource: "ValidatePackage", ofType: "sh")
+        let task = Process()
+        task.launchPath = pathToPackageValidator
+        task.arguments = [path]
+        task.launch()
+        task.waitUntilExit()
+        let status = task.terminationStatus
+        if status == 0 {
+            return true
+        }
+
+        return false
+    }
+
+    /// Validator for found scripts.
+    ///
+    /// - Parameter path: String path for file found.
+    /// - Returns: True when file is executable, has root as owner, is readable and has posix permissions of 755.
+    private func validateScript(path: String) -> Bool {
+        let fileManager = FileManager.default
+        do {
+            // Get the permissions set.
+            let itemAttributes = try fileManager.attributesOfItem(atPath: path)
+
+            // If we find the owner and posixPermissions we move forward.
+            if let owner = itemAttributes[FileAttributeKey.ownerAccountName] as? String,
+                let posix = itemAttributes[FileAttributeKey.posixPermissions] as? NSNumber {
+
+                // Convert posix Octal value to readable string
+                let posixString = String(format: "%o", posix.int16Value)
+                let isExecutable = fileManager.isExecutableFile(atPath: path)
+                let isReadable = fileManager.isReadableFile(atPath: path)
+
+                // Validate the attributes.
+                if owner == "root" && isExecutable && isReadable &&  posixString == "755"{
+                    return true
+                }
+            }
+        } catch {
+            log(message: "Script not valid: \(path)")
+        }
+
+        return false
     }
 }