+++ Source/App/Shredder/ViewControllers/ValidationViewController.swift
@@ -3,7 +3,7 @@
 //  Shredder
 //
 //  Created by Arnold Nefkens on 25/09/2018.
-//  Copyright © 2018 Pro Warehouse.
+//  Copyright © 2019 Pro Warehouse.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -83,7 +83,8 @@ class ValidationViewController: NSViewController, Logging {
                 context.allowsImplicitAnimation = true
 
                 self.quitButton.isHidden = false
-                self.continueLeadingConstraint.animator().constant = 86.0
+                let constantValue = ContinueButtonLeading.normalConstraintConstant.rawValue
+                self.continueLeadingConstraint.animator().constant = constantValue
             }, completionHandler: {
                 self.executeChecks()
             })
@@ -142,11 +143,12 @@ class ValidationViewController: NSViewController, Logging {
         validationContainer.isHidden = true
     }
 
-    private func prepareAndExecuteCommand() {
+    private func prepareAndExecuteStartOSInstallCommand() {
         log(message: "Validation: Starting Command to send")
         if let installerSelected = selectedInstaller {
 
-            let arguments: [String] = ["--agreetolicense", "--eraseinstall"]
+            let arguments: [String] = validationHelper.fetchStartOSInstallArguments()
+            log(message: "startos arguments: \(arguments)")
             var limit: Double = 300.0
             if installerSelected.version >= "14.0" {
                 limit = 600.0
@@ -154,7 +156,8 @@ class ValidationViewController: NSViewController, Logging {
 
             let completeCommandPath = "\(installerSelected.path)/Contents/Resources/startosinstall"
             let command = CommandToSend(launchPath: completeCommandPath,
-                                        launchArguments: arguments)
+                                        launchArguments: arguments,
+                                        usePTY: true)
 //            let command = CommandToSend(launchPath: "/usr/sbin/system_profiler", launchArguments: [])
             self.xpcClient.sendCommandToHelper(command: command)
             self.continueButton.isHidden = true
@@ -223,7 +226,7 @@ class ValidationViewController: NSViewController, Logging {
     }
 
     private func validateBattery() {
-        let batteryCapacityTreshold = 50
+        let batteryCapacityTreshold = 60
         if validationHelper.isRunningOnBattery(), validationHelper.currentBattryLevel() < batteryCapacityTreshold {
             log(message: "Validation: Battery Status: Fail")
             displayBatteryErrorAlert()
@@ -249,7 +252,7 @@ class ValidationViewController: NSViewController, Logging {
 
         let alert = customAlert(messageTitle: errorTitle, buttonLabels: [quitButtonTitle])
         alert.informativeText = errorMessage
-        alert.beginSheetModal(for: self.view.window!) { (response) in
+        alert.beginSheetModal(for: self.view.window!) { _ in
             self.moveToQuitApp()
         }
     }
@@ -260,14 +263,14 @@ class ValidationViewController: NSViewController, Logging {
         let newLines = "\n\n"
         if hasNoInstallerFound {
             errorMessage = NSLocalizedString("noInstallerFoundErrorMessageKey",
-                                             comment: "No Installer Found")
+                                             comment: "No Installer Found") + newLines
         }
         if hasNoNetwork {
-            errorMessage += newLines + NSLocalizedString("noNetworkMessageKey",
-                                                         comment: "No Network")
+            errorMessage += NSLocalizedString("noNetworkMessageKey",
+                                                         comment: "No Network") + newLines
         }
         if hasFindMyMacEnabled {
-            errorMessage += newLines + NSLocalizedString("findMyMacEnabledMessageKey",
+            errorMessage += NSLocalizedString("findMyMacEnabledMessageKey",
                                                          comment: "Find My Mac Enabled")
         }
 
@@ -276,7 +279,7 @@ class ValidationViewController: NSViewController, Logging {
                                 buttonLabels: [NSLocalizedString("quitKey",
                                                                  comment: "Quit")])
         alert.informativeText = errorMessage
-        alert.beginSheetModal(for: self.view.window!) { (response) in
+        alert.beginSheetModal(for: self.view.window!) { _ in
             self.moveToQuitApp()
         }
     }
@@ -301,7 +304,7 @@ class ValidationViewController: NSViewController, Logging {
                                 buttonLabels: buttonLabels)
         alert.informativeText = NSLocalizedString("messeageInstallHelperAlertKey",
                                                   comment: "Install Updated Helper")
-        alert.beginSheetModal(for: self.view.window!) { (response) in
+        alert.beginSheetModal(for: self.view.window!) { _ in
             self.xpcClient.installHelperDaemon()
         }
     }
@@ -329,14 +332,14 @@ extension ValidationViewController: XPCServiceClientProtocol {
 
     func connectionLostWhileInstallingHelper() {
         log(message: "Validation: Helper Connection lost.")
-        prepareAndExecuteCommand()
+        prepareAndExecuteStartOSInstallCommand()
     }
 
     func didReceiveVersion(value: String) {
         let localVersion = xpcClient.versionOfInstalledDeamon()
         if value == localVersion {
             log(message: "Validation: Helper versions do match")
-            self.prepareAndExecuteCommand()
+            self.prepareAndExecuteStartOSInstallCommand()
         } else {
             log(message: "Validation: Helper versions do not match")
             self.displayInstallUodatedVersionOfHelperAlert()
@@ -345,7 +348,7 @@ extension ValidationViewController: XPCServiceClientProtocol {
 
     func didInstallHelper() {
         log(message: "Validation: Helper Installed")
-        prepareAndExecuteCommand()
+        prepareAndExecuteStartOSInstallCommand()
     }
 }
 
@@ -363,7 +366,7 @@ extension ValidationViewController: ValidationCheckListViewProtocol {
 extension ValidationViewController: PickInstallerViewControllerProtocol {
     func didAnimateButtons() {
         quitButton.isHidden = true
-        continueLeadingConstraint.constant = 17.0
+        continueLeadingConstraint.constant = ContinueButtonLeading.animationConstraintConstant.rawValue
     }
 }